backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "dental-ai-backend"
version = "0.1.0"
description = "Backend GraphQL para analisis de radiografias dentales con IA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "strawberry-graphql[fastapi]>=0.275.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "motor>=3.6.0",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "PyJWT>=2.9.0",
  "ultralytics>=8.4.32",
  "pillow>=10.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
